Tests and docs

+++ modules/dsp/chowdsp_dsp_utils/Resampling/chowdsp_Downsampler.h
@@ -131,13 +131,14 @@ private:
  * using SIMD-accelerated anti-aliasing filters.
  * Note that the downsampler must be used to process block sizes
  * that are an integer multiple of the downsampling ratio.
- * Note that this requires more latency than the non-SIMD
- * version of this code, especially for higher-order filters.
+ * Note that this implementation adds num_filters * (v_size - 1)
+ * samples of latency on top of the non-SIMD
+ * implementation's latency (at the higher sample rate).
  *
  * @tparam T                        Data type to process
  * @tparam AntiAliasingFilterType   Filter type to use for an anti-aliasing filter, for example chowdsp::ButterworthFilter<8>
  */
-template <typename T, typename AntiAliasingFilterType, int maxNumChannels = defaultChannelCount>
+template <typename T, typename AntiAliasingFilterType, size_t maxNumChannels = defaultChannelCount>
 class SIMDDownsampler
 {
 public:
@@ -160,7 +161,7 @@ public:
         alignas (SIMDUtils::defaultSIMDAlignment) std::array<T, v_size> b2 {};
         for (size_t i = 0; i < num_filters; ++i)
         {
-            for (int j = 0; j < v_size; ++j)
+            for (size_t j = 0; j < v_size; ++j)
             {
                 a1[j] = aaFilter.secondOrderSections[i * v_size + j].a[1];
                 a2[j] = aaFilter.secondOrderSections[i * v_size + j].a[2];
@@ -246,11 +247,11 @@ public:
 private:
     void processFilter (const T* data_in, T* data_out, const int channel, const int numSamples) noexcept
     {
-        for (int i = 0; i < num_filters; ++i)
+        for (size_t i = 0; i < num_filters; ++i)
         {
             auto& sos = filter.secondOrderSections[i];
-            auto z1 = sos.z[channel][1];
-            auto z2 = sos.z[channel][2];
+            auto z1 = sos.z[(size_t) channel][1];
+            auto z2 = sos.z[(size_t) channel][2];
             auto x_simd_ = x_simd[(size_t) channel][i];
 
             for (int n = 0; n < numSamples; ++n)
@@ -266,8 +267,8 @@ private:
                     x_simd_[j] = x_simd_[j - 1];
             }
 
-            sos.z[channel][1] = z1;
-            sos.z[channel][2] = z2;
+            sos.z[(size_t) channel][1] = z1;
+            sos.z[(size_t) channel][2] = z2;
             x_simd[(size_t) channel][i] = x_simd_;
         }
     }
